+++ src/main/java/org/example/scheduleapiv2/user/service/UserService.java
@@ -1,13 +1,18 @@
 package org.example.scheduleapiv2.user.service;
 
 import lombok.RequiredArgsConstructor;
+import org.example.scheduleapiv2.common.util.SessionUtils;
 import org.example.scheduleapiv2.user.dto.UserCreateRequest;
+import org.example.scheduleapiv2.user.dto.UserLoginResponse;
 import org.example.scheduleapiv2.user.dto.UserResponse;
 import org.example.scheduleapiv2.user.dto.UserUpdateRequest;
 import org.example.scheduleapiv2.user.entity.User;
 import org.example.scheduleapiv2.user.repository.UserRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.ObjectUtils;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -41,15 +46,32 @@ public class UserService {
     }
 
     @Transactional
-    public UserResponse updateUser(Long userId, UserUpdateRequest request) {
+    public UserResponse updateUser(Long sessionUserId, Long userId, UserUpdateRequest request) {
         User user = userRepository.findByIdOrElseThrow(userId);
+
+        SessionUtils.assertUserIsOwner(sessionUserId, user.getId());
+
         user.updateNameAndEmail(request.getName(), request.getEmail());
 
         return UserResponse.of(user);
     }
 
-    public void deleteUser(Long userId) {
+    @Transactional
+    public void deleteUser(Long sessionUserId, Long userId) {
         User user = userRepository.findByIdOrElseThrow(userId);
+
+        SessionUtils.assertUserIsOwner(sessionUserId, user.getId());
+
         userRepository.delete(user);
+    }
+
+    public UserLoginResponse login(String email, String password) {
+        User user = userRepository.findByEmailOrElseThrow(email);
+
+        if (!ObjectUtils.nullSafeEquals(password, user.getPassword())) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Password does not match.");
+        }
+
+        return new UserLoginResponse(user.getId());
     }
 }
